Moving code to common
