+++ _common/include/Player.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <glm/glm.hpp>
+#include <glm/gtx/string_cast.hpp>
+#include "Entity.hpp"
+#include "Mover.hpp"
+
+class Player : public Entity {
+public:
+    // glm::vec2 inputs = glm::vec2();
+    // glm::vec3 velocity = glm::vec3();
+    // float speed = 0.2f;
+
+    Player();
+
+    virtual void update() override;
+    int32_t TypeID() const override;
+};
