+++ client/include/Player.h
@@ -5,7 +5,7 @@
 #include "Entity.h"
 
 class Player : public Entity {
-    
+    Player();
 };
