+++ client/include/PlayerManager.hpp
@@ -1,11 +1,14 @@
 #ifndef PLAYER_MANAGER_H
 #define PLAYER_MANAGER_H
 #include <string>
-#include "Mover.h"
+#include "Player.h"
+
 class PlayerManager {
 public:
     int id;
     std::string username;
-    Mover* mover;
+    // Mover* mover;
+
+    Player* player;
 };
 #endif
