+++ client/src/Mover.cpp
@@ -1,34 +1,43 @@
 #include "Mover.h"
+#include "NetTransform.hpp"
 
-Mover::Mover() : INetworkComponent() {
-    position = glm::vec3(0);
-    velocityHeading = glm::vec3(0);
-    speed = 0.2;
+Mover::Mover(NetworkObject* owner) : 
+    INetworkComponent(owner),
+    position(owner->GetComponent<NetTransform>()->position)
+{
+    // position = glm::vec3(0);
+    // velocityHeading = glm::vec3(0);
+    // speed = 0.2;
 }
 
-Mover::Mover(std::string path) : INetworkComponent() {
-    position = glm::vec3(0);
-    velocityHeading = glm::vec3(0);
-    speed = 0.2;
+Mover::Mover(NetworkObject* owner, std::string path) : 
+    INetworkComponent(owner),
+    position(owner->GetComponent<NetTransform>()->position)
+{
+    // position = glm::vec3(0);
+    // velocityHeading = glm::vec3(0);
+    // speed = 0.2; 
+
     model = new Model(path);
     clip = new AnimationClip(path, model);
     animationPlayer = new AnimationPlayer(clip);
+    // position = owner->GetComponent<NetTransform>()->position;
 }
 
 Mover::~Mover() {}
 
-void Mover::Update(float deltaTime) {
-    model->update(deltaTime, position);
-    animationPlayer->update(deltaTime);
-}
+// void Mover::Update(float deltaTime) {
+//     model->update(deltaTime, position);
+//     animationPlayer->update(deltaTime);
+// }
 
-void Mover::UpdatePhysics(float deltaTime) {
-    glm::vec3 normalizedVelocityHeading = velocityHeading;
-    if (velocityHeading != glm::vec3(0)) {
-        normalizedVelocityHeading = glm::normalize(velocityHeading);
-    }
-    position += normalizedVelocityHeading * speed;
-}
+// void Mover::UpdatePhysics(float deltaTime) {
+//     glm::vec3 normalizedVelocityHeading = velocityHeading;
+//     if (velocityHeading != glm::vec3(0)) {
+//         normalizedVelocityHeading = glm::normalize(velocityHeading);
+//     }
+//     position += normalizedVelocityHeading * speed;
+// }
 
 void Mover::Draw(glm::mat4 view, GLuint shaderProgram) {
     // glm::mat4 transform = glm::translate(position);
