+++ server/src/NetworkManager.cpp
@@ -83,7 +83,7 @@ void NetworkManager::process_input() {
         // but for now, we do this to set input manually
         switch ((PacketType)packet_type) {
         case PacketType::PLAYER_INPUT: {
-            std::cout << "  RECV: PINPUT" << std::endl;
+            // std::cout << "  RECV: PINPUT" << std::endl;
             char input[4];
             for (int i = 0; i < 4; i++) {
                 packet->read_byte(&input[i]);
@@ -93,7 +93,7 @@ void NetworkManager::process_input() {
             if (clients.find(client_id) == clients.end())
                 break;
 
-            std::cout << "  Received input: " << (float)input[0] << ", " << (float)input[1] << ", " << (float)input[2] << ", " << (float)input[3] << std::endl;
+            // std::cout << "  Received input: " << (float)input[0] << ", " << (float)input[1] << ", " << (float)input[2] << ", " << (float)input[3] << std::endl;
 
             clients[client_id]->p->GetComponent<Mover>()->input.x = (float)input[3] - (float)input[1];
             clients[client_id]->p->GetComponent<Mover>()->input.y = (float)input[0] - (float)input[2];
@@ -101,7 +101,7 @@ void NetworkManager::process_input() {
             break;
         }
         case PacketType::DESTROY_OBJECT_ACK:
-            std::cout << "  RECV: DSTRY_ACK" << std::endl;
+            // std::cout << "  RECV: DSTRY_ACK" << std::endl;
             int numObjectsDestroyed;
             packet->read_int(&numObjectsDestroyed);
             while (numObjectsDestroyed) {
@@ -116,7 +116,7 @@ void NetworkManager::process_input() {
             }
             break;
         default:
-            std::cout << "  RECV: NONEOFTHEABOVE" << std::endl;
+            // std::cout << "  RECV: NONEOFTHEABOVE" << std::endl;
             break;
         }
     }
